최종 수정본 3

+++ Application/main.cpp
@@ -17,7 +17,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance,
 	_In_ LPWSTR    lpCmdLine,
 	_In_ int       nCmdShow)
 {
-	srand(time(nullptr));
+	srand(static_cast<unsigned int>(time(nullptr)));
 	PuyoPuyoContentsCore User;
 	return UEngineAPICore::EngineStart(hInstance, &User);
 }
